pyproject.toml
[project]
name = "oauth2-server"
version = "1.0.0"
description = "OAuth2 Authorization Server built with FastAPI"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.6",
    "uvicorn[standard]>=0.34.0",
    "sqlalchemy[asyncio]>=2.0.36",
    "aiomysql>=0.2.0",
    "alembic>=1.14.1",
    "python-jose[cryptography]>=3.3.0",
    "passlib>=1.7.4",
    "bcrypt==3.2.0",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.5",
    "itsdangerous>=2.2.0",
    "pydantic[email]>=2.10.4",
    "pydantic-settings>=2.7.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.0",
    "httpx>=0.28.1",
    "aiosqlite>=0.20.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
